Move the steps routes to a distinct controller and fix flashes
* move to diagnoses/steps
* rename the GET “stepX” routes to the same routes as the POST routes
* move the partial and rename the i18n keys

This should make things easier for #791 and subsequent redesign of the diagnosis steps.

Additional related changes:

* Define named constants for the 2, 3, 4 diagnosis steps. This is not yet a proper enum, but we’re getting there.
* Fix flash messages not being displayed after remote forms; we still need to use flashes.js.haml, but at least it’s working.
