+++ app/assets/stylesheets/application/diagnoses.sass
@@ -1,4 +1,4 @@
-#step2-app
+#besoins-app
   .subject.checkbox
     display: block
   .subject
@@ -11,7 +11,7 @@
     input[type=checkbox]:checked ~ textarea
       display: block
 
-#step4-app
+#selection-app
   td.checkbox-width
     width: 10px
   td.expert-width
