+++ app/views/diagnoses/steps/_expert_subject_checkboxes.html.haml
@@ -14,11 +14,11 @@
       %td
         = expert_subject_form.label checkbox_id, class: 'clickable' do
           .tiny.text.bold
-            = t('diagnoses.step4.expert_scope_of_work')
+            = t('diagnoses.steps.selection.expert_scope_of_work')
           .small.text
             = expert_subject.institution_subject.description
           - if expert_subject.description.present?
             .tiny.text.bold
-              = t('diagnoses.step4.more_informations')
+              = t('diagnoses.steps.selection.more_informations')
             .small.text
               = expert_subject.description
