+++ app/views/shared/flashes.js.haml
@@ -1,4 +1,4 @@
-var flashes = "#{j render('layouts/flashes')}";
+var flashes = "#{j render('flashes')}";
 document.getElementById('flashes').innerHTML = flashes;
 
 :ruby
