+++ config/i18n-tasks.yml
@@ -92,7 +92,7 @@ ignore_unused:
   - 'active_admin.scopes.*'
   - 'active_admin.move'
   - 'attributes.*'
-  - 'diagnoses.steps.*'
+  - 'diagnoses.steps.*.title'
   - 'mailers.admin_mailer.weekly_statistics.needs_abandoned.scopes.*'
   - 'mailers.user_mailer.update_match_notify.expert_and_status.*'
   - 'stats.stats_table.*'
