+++ spec/views/diagnoses/steps/step4.html.haml_spec.rb
@@ -1,7 +1,7 @@
 require "spec_helper"
 require 'rails_helper'
 
-describe "diagnoses/step4.html.haml", type: :view do
+describe "diagnoses/steps/selection.html.haml", type: :view do
   let(:need) { create :need }
   let(:diagnosis) { create :diagnosis, needs: [need] }
   let(:institution_subject) { create :institution_subject, subject: need.subject }
